Fixed imports and api calls in ceph.py. Also moved all clustering code from metric_tracking.__init__.py to clustering.__init__.py
