+++ storage/ceph.py
@@ -1,5 +1,5 @@
 import pyarrow.parquet as pq
-import s3fs, datetime, os, s3, local
+import s3fs, datetime, os
 
 (ceph_key, ceph_secret, ceph_endpoint) = (os.getenv('CEPH_KEY'), os.getenv('CEPH_SECRET'), os.getenv('CEPH_ENDPOINT'))
 client_kwargs = {'endpoint_url': ceph_endpoint}
@@ -10,13 +10,10 @@ def get_dataset(key):
     return pq.ParquetDataset(key, filesystem=filesystem).read_pandas().to_pandas()
 
 
-def write(data, use_local=True):
+def write(data):
 	date = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-	if use_local:
-		local.write(data, date)
-	else:
-	    s3.write(data, date)
+	write(data, date)
 
 
-def read(date, use_local=True):
-	raise NotImplementedError
+def read(date):
+	return read(date=date)
