+++ sync.py
@@ -4,7 +4,6 @@ logging.basicConfig(level=logging.DEBUG)
 
 import clustering
 import storage
-import stability_score as ss
 
 def sync():
     for key, date in storage.unprocessed():
